fix: solucionado error de falta de filtro por usario de espacios
