+++ app/spaces/page.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { auth } from "@/auth"
 import { redirect } from "next/navigation";
 import AddSpaceForm from "@/components/spaces/AddSpaceForm";
@@ -6,9 +8,47 @@ import SpaceList from "@/components/spaces/SpaceList";
 import BtnSignOut from "@/components/signout";
 import Link from "next/link";
 import { getUserAuthenticated } from "../../services/authService";
+import { useCallback, useEffect, useState } from "react";
+import { supabase } from "@/lib/supabaseClient";
+import { useDispatch } from "react-redux";
+import { RootState } from "@/redux/store";
+import { useSelector } from "react-redux";
+import { updateInitialState } from '@/redux/slices/spaceSlice';
 
-async function Spaces() {
-  const session = await auth()
+
+function Spaces() {
+  const [error, setError] = useState<string>()
+  const dispatch = useDispatch()
+  const user = useSelector((state : RootState) => {
+    return state.user
+  })
+  const spaces = useSelector((state : RootState) => {
+    return state.spaces
+  })
+
+  const getSpaces = useCallback((async () => {
+    const { data , error } = await supabase.from('spaces').select('*').gte('user_id', `${user.user?.id}` )
+
+    if(error) {
+      console.log(error)
+      setError(error.message)
+    }
+
+    if(data){
+      dispatch(updateInitialState(data)) 
+    }
+
+    if(data?.length === 0) {
+      return (
+        <>No hay espacios</>
+      )
+    }
+  }),[dispatch, user.user?.id]
+)
+
+  useEffect(() => {
+    getSpaces()
+  }, [getSpaces])
   
 
 
@@ -25,14 +65,15 @@ async function Spaces() {
       </div>
       <main>
         {
-          session &&
             <SessionProvider>
               <div className="flex flex-col gap-4 p-4">
                 <header>
                   <AddSpaceForm />
                 </header>
                 <section className="border-luissdev-250 shadow-md shadow-luissdev-250 py-4 px-2 rounded-md bg-luissdev-250">
-                  <SpaceList />
+                  <SpaceList 
+                    spaces={ spaces }
+                  />
                 </section>
               </div>
             </SessionProvider>
