+++ services/spaceService.ts
@@ -2,8 +2,8 @@ import { supabase } from "@/lib/supabaseClient";
 import { ISpace } from "@/types/types";
 
 
-export async function getAllSpaces(): Promise<ISpace[]> {
-  const { data, error } = await supabase.from('spaces').select('*') 
+export async function getAllSpaces(userId : string): Promise<ISpace[]> {
+  const { data, error } = await supabase.from('spaces').select('*').gte('user_id',`${userId}`) 
   if(error) {
     console.log(error)
     throw new Error('Error al obtener los espacios')
